refactor: qc-application payment 도메인 패키지 정리
