+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/point/implement/ChargePointPaymentFailHandler.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.payment;
+package com.eager.questioncloud.application.api.payment.point.implement;
 
 
 import com.eager.questioncloud.core.domain.point.infrastructure.ChargePointPaymentRepository;
