+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/point/implement/PGAPI.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.payment;
+package com.eager.questioncloud.application.api.payment.point.implement;
 
 
 import com.eager.questioncloud.core.domain.point.dto.PGPayment;
