+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/question/controller/QuestionPaymentController.java
@@ -1,8 +1,10 @@
-package com.eager.questioncloud.application.api.payment;
+package com.eager.questioncloud.application.api.payment.question.controller;
 
 import com.eager.questioncloud.application.api.common.DefaultResponse;
 import com.eager.questioncloud.application.api.common.PagingResponse;
-import com.eager.questioncloud.application.api.payment.QuestionPaymentControllerRequest.QuestionPaymentRequest;
+import com.eager.questioncloud.application.api.payment.question.dto.QuestionPaymentControllerRequest.QuestionPaymentRequest;
+import com.eager.questioncloud.application.api.payment.question.service.QuestionPaymentHistoryService;
+import com.eager.questioncloud.application.api.payment.question.service.QuestionPaymentService;
 import com.eager.questioncloud.application.security.UserPrincipal;
 import com.eager.questioncloud.core.common.PagingInformation;
 import com.eager.questioncloud.core.domain.payment.dto.QuestionPaymentHistory;
