+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/question/service/QuestionPaymentService.java
@@ -1,5 +1,7 @@
-package com.eager.questioncloud.application.api.payment;
+package com.eager.questioncloud.application.api.payment.question.service;
 
+import com.eager.questioncloud.application.api.payment.question.implement.QuestionPaymentGenerator;
+import com.eager.questioncloud.application.api.payment.question.implement.QuestionPaymentProcessor;
 import com.eager.questioncloud.core.domain.payment.event.CompletedQuestionPaymentEvent;
 import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
 import com.eager.questioncloud.lock.LockKeyGenerator;
